backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
git+https://github.com/sentient-agi/ROMA.git
openai==1.0.0
